Rename EntitySelect → EntityControl
And pull the component out in a file.
